fix refresh animation
while trying out a fully css & :active solution, this approach made the
button react oddly, starting or ending the animation in a not-so-smooth
manner.

the :active pseudo was providing me a mouseup and mousedown animation
which was unwanted. therefore, i just ended up implementing the desired
effect in jQuery - and saved the mark up ~10 lines of code.

Not ideal, but I tried. jQuery wins